ListboxTextItem does not use the custom font in parseTextString()

+++ samples/browser/include/SampleBrowser.h
@@ -66,7 +66,7 @@ public:
 
     void update(float passedTime) override;
 
-    void handleNewWindowSize(float width, float height) override;
+    void handleNewWindowSize(int width, int height) override;
 
     void renderGUIContexts() override;
